No commit message
svn path=/trunk/iCub/firmware/; revision=22786

+++ emBODY/arch-armcmx/body/abslayer/hal-2012-package/hal/hal-src/extra/periphs/hal_periph_timer_hid.h
@@ -47,7 +47,7 @@
 
 typedef struct
 {
-    uint8_t                 supported_mask;
+    uint16_t            supported_mask;
 } hal_timer_hid_brdcfg_t;
 
 // - declaration of extern hidden variables ---------------------------------------------------------------------------
